(0.3 * score + 0.7 * time_over / 105) * size_score / 5;

+++ disk.cpp
@@ -194,7 +194,7 @@ void Disk::reflash_partition_score(){
     }
     initialize_partitions();
 }
-void Disk::update_partition_info(int partition_id, float score){
+void Disk::update_partition_info(int partition_id, double score){
     // if (partition_id <= 0 || partition_id > partitions.size() - 1 || score <= 0) return;
     partitions[partition_id].score = score;
     // 调用动态堆 update 操作，调整该分区在堆中的位置
